refactor: expose the OrderBy type as part of the public api
