+++ firebase-firestore/src/main/java/com/google/firebase/firestore/core/Bound.java
@@ -16,7 +16,8 @@ package com.google.firebase.firestore.core;
 
 import static com.google.firebase.firestore.util.Assert.hardAssert;
 
-import com.google.firebase.firestore.core.OrderBy.Direction;
+import com.google.firebase.firestore.OrderBy;
+import com.google.firebase.firestore.OrderBy.Direction;
 import com.google.firebase.firestore.model.Document;
 import com.google.firebase.firestore.model.DocumentKey;
 import com.google.firebase.firestore.model.FieldPath;
@@ -85,7 +86,7 @@ public final class Bound {
     for (int i = 0; i < position.size(); i++) {
       OrderBy orderByComponent = orderBy.get(i);
       FieldValue component = position.get(i);
-      if (orderByComponent.field.equals(FieldPath.KEY_PATH)) {
+      if (orderByComponent.getField().equals(FieldPath.KEY_PATH)) {
         Object refValue = component.value();
         hardAssert(
             refValue instanceof DocumentKey,
