+++ firebase-firestore/src/main/java/com/google/firebase/firestore/remote/RemoteSerializer.java
@@ -25,8 +25,8 @@ import com.google.firebase.firestore.core.Bound;
 import com.google.firebase.firestore.core.Filter;
 import com.google.firebase.firestore.core.NaNFilter;
 import com.google.firebase.firestore.core.NullFilter;
-import com.google.firebase.firestore.core.OrderBy;
-import com.google.firebase.firestore.core.OrderBy.Direction;
+import com.google.firebase.firestore.OrderBy;
+import com.google.firebase.firestore.OrderBy.Direction;
 import com.google.firebase.firestore.core.Query;
 import com.google.firebase.firestore.core.RelationFilter;
 import com.google.firebase.firestore.local.QueryData;
