+++ firebase-firestore/src/testUtil/java/com/google/firebase/firestore/testutil/TestUtil.java
@@ -35,8 +35,8 @@ import com.google.firebase.firestore.TestAccessHelper;
 import com.google.firebase.firestore.UserDataConverter;
 import com.google.firebase.firestore.core.Filter;
 import com.google.firebase.firestore.core.Filter.Operator;
-import com.google.firebase.firestore.core.OrderBy;
-import com.google.firebase.firestore.core.OrderBy.Direction;
+import com.google.firebase.firestore.OrderBy;
+import com.google.firebase.firestore.OrderBy.Direction;
 import com.google.firebase.firestore.core.Query;
 import com.google.firebase.firestore.core.UserData.ParsedUpdateData;
 import com.google.firebase.firestore.local.LocalViewChanges;
